update baseline so we do not depend on any detached plugins
As this plugin depended on detached plugins installing it from the
outset in the ATH caused the trilead-api plugin to be installed, and
this could be an older version that was required by some plugins that we
actually want to test.  This causes an issue as the plugins are not hot
installable as the update to trilead-api requires a restart - which does
not happen and the ATH fails.

Observed in the Git plugin tests.